kafka/requirements.txt
asyncio==3.4.3
bidict==0.22.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.2.5
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
numpy==1.21.6
pandas==1.3.5
py4j==0.10.9.7
pyspark==3.4.1
python-dateutil==2.8.2
python-engineio==4.7.1
python-socketio==4.5.0
pytz==2023.3.post1
requests==2.31.0
simple-websocket==0.10.1
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
websocket-client==1.6.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
